pyproject.toml
[project]
name = "cua"
version = "0.1.0"
description = "CUA using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.12.2",
    "faster-whisper>=1.2.1",
    "yt-dlp>=2026.3.17",
    "python-socketio>=4.6,<5",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "langchain",
    "langchain-community",
    "langchain-core",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "langchain-openai>=0.0.6",
    "langchain-groq>=0.0.1",
    "reportlab>=4.0.0",
    "fpdf2>=2.8.7",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "requests>=2.31.0",
]

[project.scripts]
cua = "cua.main:run"
run_crew = "cua.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
